Changed to fillPage and other stuff
In this commit, the bigger container is changed from fixedPage to fillPage, and a footpage was added with a link to the source code on github. In addition, now when the Play button is clicked, the number in the textinput (previously a numericInput) becomes a placeholder and the textinput value is emptied.